# M2-Challenge-Wheat-Rayce/src/main/java/com/example/M2ChallengeWheatRayce/models/CustomErrorResponse.java
package com.example.M2ChallengeWheatRayce.models;


import com.fasterxml.jackson.annotation.JsonFormat;

import java.time.LocalDateTime;

//Copied from Boilerplate form class
public class CustomErrorResponse {

    String errorMsg;

    int status;

    String errorCode;

    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd hh:mm:ss")
    LocalDateTime timestamp;

    public CustomErrorResponse(){}

    public CustomErrorResponse(String errorCode, String errorMsg){
        this.errorCode = errorCode;
        this.errorMsg = errorMsg;
    }

    public String getErrorMsg() {
        return errorMsg;
    }

    public void setErrorMsg(String errorMsg) {
        this.errorMsg = errorMsg;
    }

    public int getStatus() {
        return status;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public String getErrorCode() {
        return errorCode;
    }

    public void setErrorCode(String errorCode) {
        this.errorCode = errorCode;
    }

    public LocalDateTime getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(LocalDateTime timestamp) {
        this.timestamp = timestamp;
    }
}