# M2-Challenge-Wheat-Rayce/src/main/java/com/example/M2ChallengeWheatRayce/models/MathSolution.java
package com.example.M2ChallengeWheatRayce.models;

public class MathSolution {

    private int operand1;

    private int operand2;

    private String operation;

    private int answer;

    public int getOperand1() {
        return operand1;
    }

    public void setOperand1(int operand1) {
        this.operand1 = operand1;
    }

    public int getOperand2() {
        return operand2;
    }

    public void setOperand2(int operand2) {
        this.operand2 = operand2;
    }

    public String getOperation() {
        return operation;
    }

    public void setOperation(String operation) {
        this.operation = operation;
    }

    public int getAnswer() {
        return answer;
    }

    public void setAnswer(int answer) {
        this.answer = answer;
    }

    public MathSolution() {

    }

    public MathSolution(int operand1, int operand2, String operation, int answer) {
        this.operand1 = operand1;
        this.operand2 = operand2;
        this.operation = operation;
        this.answer = answer;
    }
}